+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added
   - Transform `P` type argument
+  - Transform `setState` (simple case)
 
 ## 0.1.0